+++ Packages/com.unity.input/README.md
@@ -1,3 +1,3 @@
 A new Input System for Unity.
 
-Check out the [Input System documentation](http://docs.unity3d.com/Packages/com.unity.inputsystem@latest/) for more info.
+Check out the [Input System documentation](http://docs.unity3d.com/Packages/com.unity.input@latest/) for more info.
