+++ Packages/com.unity.input/Runtime/Actions/Interactions/HoldInteraction.cs
@@ -130,7 +130,7 @@ namespace UnityEngine.InputSystem.Interactions
 
         public override void OnDrawVisualElements(VisualElement root, Action onChangedCallback)
         {
-	        m_PressPointSetting.OnDrawVisualElements(root, onChangedCallback);
+            m_PressPointSetting.OnDrawVisualElements(root, onChangedCallback);
             m_DurationSetting.OnDrawVisualElements(root, onChangedCallback);
         }
