+++ Packages/com.unity.input/Runtime/Actions/Interactions/SlowTapInteraction.cs
@@ -94,8 +94,8 @@ namespace UnityEngine.InputSystem.Interactions
 
         public override void OnDrawVisualElements(VisualElement root, Action onChangedCallback)
         {
-	        m_DurationSetting.OnDrawVisualElements(root, onChangedCallback);
-	        m_PressPointSetting.OnDrawVisualElements(root, onChangedCallback);
+            m_DurationSetting.OnDrawVisualElements(root, onChangedCallback);
+            m_PressPointSetting.OnDrawVisualElements(root, onChangedCallback);
         }
 
         private CustomOrDefaultSetting m_DurationSetting;
