+++ Packages/com.unity.input/Runtime/Actions/Interactions/TapInteraction.cs
@@ -108,7 +108,7 @@ namespace UnityEngine.InputSystem.Interactions
 
         public override void OnDrawVisualElements(VisualElement root, Action onChangedCallback)
         {
-	        m_DurationSetting.OnDrawVisualElements(root, onChangedCallback);
+            m_DurationSetting.OnDrawVisualElements(root, onChangedCallback);
             m_PressPointSetting.OnDrawVisualElements(root, onChangedCallback);
         }
