+++ Packages/com.unity.input/Runtime/AssemblyInfo.cs
@@ -7,16 +7,15 @@ using UnityEngine.InputSystem;
 [assembly: InternalsVisibleTo("Unity.InputSystem.Tests.Editor")]
 [assembly: InternalsVisibleTo("Unity.InputSystem.Tests")]
 [assembly: InternalsVisibleTo("Unity.InputSystem.IntegrationTests")]
-[assembly: InternalsVisibleTo("UnityEngine.InputModuleLegacy")]
 
 namespace UnityEngine.InputSystem
 {
     public static partial class InputSystem
     {
-        // Keep this in sync with "Packages/com.unity.inputsystem/package.json".
+        // Keep this in sync with "Packages/com.unity.input/package.json".
         // NOTE: Unfortunately, System.Version doesn't use semantic versioning so we can't include
         //       "-preview" suffixes here.
         internal const string kAssemblyVersion = "2.0.0";
-        internal const string kDocUrl = "https://docs.unity3d.com/Packages/com.unity.inputsystem@2.0";
+        internal const string kDocUrl = "https://docs.unity3d.com/Packages/com.unity.input@2.0";
     }
 }
