+++ Packages/com.unity.input/Runtime/Editor/AssetEditor/InputActionTreeView.cs
@@ -1492,7 +1492,7 @@ namespace UnityEngine.InputSystem.Editor
         private static readonly GUIContent s_ExpandAllLabel = EditorGUIUtility.TrTextContent("Expand All");
         private static readonly GUIContent s_CollapseAllLabel = EditorGUIUtility.TrTextContent("Collapse All");
 
-        public static string SharedResourcesPath = "Packages/com.unity.inputsystem/InputSystem/Editor/AssetEditor/Resources/";
+        public static string SharedResourcesPath = "Packages/com.unity.input/Runtime/Editor/AssetEditor/Resources/";
         public static string ResourcesPath
         {
             get
