+++ Packages/com.unity.input/Runtime/Editor/AssetEditor/ParameterListView.cs
@@ -200,21 +200,21 @@ namespace UnityEngine.InputSystem.Editor.Lists
                 m_ParameterEditor = (InputParameterEditor)Activator.CreateInstance(parameterEditorType);
 
                 // We have to jump through some hoops here to create instances of any CustomOrDefaultSetting fields on the
-                // parameter editor. This is because those types changed from structs to classes when UIToolkit was 
+                // parameter editor. This is because those types changed from structs to classes when UIToolkit was
                 // introduced, and we don't want to force users to have to create those instances manually on any of their
                 // own editors.
-                var genericArgumentType = TypeHelpers.GetGenericTypeArgumentFromHierarchy(parameterEditorType, 
-	                typeof(InputParameterEditor<>), 0);
-                if(genericArgumentType != null)
+                var genericArgumentType = TypeHelpers.GetGenericTypeArgumentFromHierarchy(parameterEditorType,
+                    typeof(InputParameterEditor<>), 0);
+                if (genericArgumentType != null)
                 {
-	                var fieldInfos = parameterEditorType
-		                .GetFields(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance);
-	                var customOrDefaultGenericType = typeof(InputParameterEditor<>.CustomOrDefaultSetting);
-	                var customOrDefaultType = customOrDefaultGenericType.MakeGenericType(genericArgumentType);
-	                foreach (var customOrDefaultEditorField in fieldInfos.Where(f => f.FieldType == customOrDefaultType))
-	                {
-		                customOrDefaultEditorField.SetValue(m_ParameterEditor, Activator.CreateInstance(customOrDefaultEditorField.FieldType));
-	                }
+                    var fieldInfos = parameterEditorType
+                        .GetFields(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance);
+                    var customOrDefaultGenericType = typeof(InputParameterEditor<>.CustomOrDefaultSetting);
+                    var customOrDefaultType = customOrDefaultGenericType.MakeGenericType(genericArgumentType);
+                    foreach (var customOrDefaultEditorField in fieldInfos.Where(f => f.FieldType == customOrDefaultType))
+                    {
+                        customOrDefaultEditorField.SetValue(m_ParameterEditor, Activator.CreateInstance(customOrDefaultEditorField.FieldType));
+                    }
                 }
                 m_ParameterEditor.SetTarget(instance);
             }
@@ -248,76 +248,76 @@ namespace UnityEngine.InputSystem.Editor.Lists
 
         public void OnDrawVisualElements(VisualElement root)
         {
-	        if (m_ParameterEditor != null)
-	        {
+            if (m_ParameterEditor != null)
+            {
                 m_ParameterEditor.OnDrawVisualElements(root, OnValuesChanged);
-		        return;
-	        }
+                return;
+            }
 
-	        if (m_Parameters == null)
-	            return;
+            if (m_Parameters == null)
+                return;
 
             void OnValueChanged(ref EditableParameterValue parameter, object result, int i)
             {
-	            parameter.value.value = PrimitiveValue.FromObject(result).ConvertTo(parameter.value.type);
-	            m_Parameters[i] = parameter;
-	            onChange?.Invoke();
+                parameter.value.value = PrimitiveValue.FromObject(result).ConvertTo(parameter.value.type);
+                m_Parameters[i] = parameter;
+                onChange?.Invoke();
             }
 
             for (var i = 0; i < m_Parameters.Length; i++)
             {
-	            var parameter = m_Parameters[i];
-	            var label = m_ParameterLabels[i];
-	            var closedIndex = i;
-                
-	            if (parameter.isEnum)
-	            {
-		            var intValue = parameter.value.value.ToInt32();
+                var parameter = m_Parameters[i];
+                var label = m_ParameterLabels[i];
+                var closedIndex = i;
+
+                if (parameter.isEnum)
+                {
+                    var intValue = parameter.value.value.ToInt32();
                     var field = new DropdownField(label.text, parameter.enumNames.Select(x => x.text).ToList(), intValue);
                     field.RegisterValueChangedCallback(evt => OnValueChanged(ref parameter, evt.newValue, closedIndex));
                     root.Add(field);
-	            }
-	            else if (parameter.value.type == TypeCode.Int64 || parameter.value.type == TypeCode.UInt64)
-	            {
-		            var longValue = parameter.value.value.ToInt64();
-		            var field = new LongField(label.text) { value = longValue };
+                }
+                else if (parameter.value.type == TypeCode.Int64 || parameter.value.type == TypeCode.UInt64)
+                {
+                    var longValue = parameter.value.value.ToInt64();
+                    var field = new LongField(label.text) { value = longValue };
                     field.RegisterValueChangedCallback(evt => OnValueChanged(ref parameter, evt.newValue, closedIndex));
                     root.Add(field);
-	            }
-	            else if (parameter.value.type.IsInt())
-	            {
-		            var intValue = parameter.value.value.ToInt32();
-		            var field = new IntegerField(label.text) { value = intValue };
-		            field.RegisterValueChangedCallback(evt => OnValueChanged(ref parameter, evt.newValue, closedIndex));
+                }
+                else if (parameter.value.type.IsInt())
+                {
+                    var intValue = parameter.value.value.ToInt32();
+                    var field = new IntegerField(label.text) { value = intValue };
+                    field.RegisterValueChangedCallback(evt => OnValueChanged(ref parameter, evt.newValue, closedIndex));
                     root.Add(field);
-	            }
-	            else if (parameter.value.type == TypeCode.Single)
-	            {
-		            var floatValue = parameter.value.value.ToSingle();
-		            var field = new FloatField(label.text) { value = floatValue };
-		            field.RegisterValueChangedCallback(evt => OnValueChanged(ref parameter, evt.newValue, closedIndex));
+                }
+                else if (parameter.value.type == TypeCode.Single)
+                {
+                    var floatValue = parameter.value.value.ToSingle();
+                    var field = new FloatField(label.text) { value = floatValue };
+                    field.RegisterValueChangedCallback(evt => OnValueChanged(ref parameter, evt.newValue, closedIndex));
                     root.Add(field);
-	            }
-	            else if (parameter.value.type == TypeCode.Double)
-	            {
-		            var floatValue = parameter.value.value.ToDouble();
-		            var field = new DoubleField(label.text) { value = floatValue };
-		            field.RegisterValueChangedCallback(evt => OnValueChanged(ref parameter, evt.newValue, closedIndex));
+                }
+                else if (parameter.value.type == TypeCode.Double)
+                {
+                    var floatValue = parameter.value.value.ToDouble();
+                    var field = new DoubleField(label.text) { value = floatValue };
+                    field.RegisterValueChangedCallback(evt => OnValueChanged(ref parameter, evt.newValue, closedIndex));
                     root.Add(field);
-	            }
-	            else if (parameter.value.type == TypeCode.Boolean)
-	            {
-		            var boolValue = parameter.value.value.ToBoolean();
-		            var field = new Toggle(label.text) { value = boolValue };
-		            field.RegisterValueChangedCallback(evt => OnValueChanged(ref parameter, evt.newValue, closedIndex));
+                }
+                else if (parameter.value.type == TypeCode.Boolean)
+                {
+                    var boolValue = parameter.value.value.ToBoolean();
+                    var field = new Toggle(label.text) { value = boolValue };
+                    field.RegisterValueChangedCallback(evt => OnValueChanged(ref parameter, evt.newValue, closedIndex));
                     root.Add(field);
-	            }
+                }
             }
         }
 
         private void OnValuesChanged()
         {
-	        ReadParameterValuesFrom(m_ParameterEditor.target);
+            ReadParameterValuesFrom(m_ParameterEditor.target);
             onChange?.Invoke();
         }
