+++ Packages/com.unity.input/Runtime/Editor/AssetImporter/InputActionImporter.cs
@@ -29,8 +29,8 @@ namespace UnityEngine.InputSystem.Editor
     {
         private const int kVersion = 13;
 
-        private const string kActionIcon = "Packages/com.unity.inputsystem/InputSystem/Editor/Icons/InputAction.png";
-        private const string kAssetIcon = "Packages/com.unity.inputsystem/InputSystem/Editor/Icons/InputActionAsset.png";
+        private const string kActionIcon = "Packages/com.unity.input/Runtime/Editor/Icons/InputAction.png";
+        private const string kAssetIcon = "Packages/com.unity.input/Runtime/Editor/Icons/InputActionAsset.png";
 
         [SerializeField] private bool m_GenerateWrapperCode;
         [SerializeField] private string m_WrapperCodePath;
