+++ Packages/com.unity.input/Runtime/Editor/GlobalAssetEditor/ReactiveProperty.cs
@@ -2,24 +2,24 @@ using System;
 
 namespace UnityEngine.InputSystem.Editor
 {
-	internal class ReactiveProperty<T>
-	{
-		private T m_Value;
-		public event Action<T> Changed;
+    internal class ReactiveProperty<T>
+    {
+        private T m_Value;
+        public event Action<T> Changed;
 
-		public T value
-		{
-			get => m_Value;
-			set
-			{
-				m_Value = value;
-				Changed?.Invoke(m_Value);
-			}
-		}
+        public T value
+        {
+            get => m_Value;
+            set
+            {
+                m_Value = value;
+                Changed?.Invoke(m_Value);
+            }
+        }
 
-		public void SetValueWithoutChangeNotification(T value)
-		{
-			m_Value = value;
-		}
-	}
+        public void SetValueWithoutChangeNotification(T value)
+        {
+            m_Value = value;
+        }
+    }
 }
