+++ Packages/com.unity.input/Runtime/Editor/GlobalAssetEditor/Resources/BindingPanelRowTemplate.uxml
@@ -2,7 +2,7 @@
     <ui:VisualElement name="item-row" class="unity-list-view__item" style="flex-direction: column; flex-grow: 1; justify-content: center; margin-left: 4px;">
         <ui:VisualElement name="row" style="flex-direction: row;">
             <ui:Toggle name="expando" class="unity-foldout__toggle" style="visibility: hidden;" />
-            <ui:VisualElement name="icon" style="justify-content: center; background-image: url(&apos;project://database/Packages/com.unity.inputsystem/InputSystem/Editor/Icons/d_Gamepad.png?fileID=2800000&amp;guid=76dc61ff6c2d84b24b0a2ce01dc2842d&amp;type=3#d_Gamepad&apos;); width: 16px; height: 16px;" />
+            <ui:VisualElement name="icon" style="justify-content: center; background-image: url(&apos;project://database/Packages/com.unity.input/Runtime/Editor/Icons/d_Gamepad.png?fileID=2800000&amp;guid=76dc61ff6c2d84b24b0a2ce01dc2842d&amp;type=3#d_Gamepad&apos;); width: 16px; height: 16px;" />
             <ui:Label text="binding-name" display-tooltip-when-elided="true" name="name" style="flex-grow: 1; justify-content: center; align-items: stretch; margin-left: 4px; -unity-font-style: bold;" />
         </ui:VisualElement>
     </ui:VisualElement>
