+++ Packages/com.unity.input/Runtime/Editor/GlobalAssetEditor/Resources/NameAndParameterListViewItemTemplate.uxml
@@ -1,5 +1,5 @@
 <ui:UXML xmlns:ui="UnityEngine.UIElements" xmlns:uie="UnityEditor.UIElements" xsi="http://www.w3.org/2001/XMLSchema-instance" engine="UnityEngine.UIElements" editor="UnityEditor.UIElements" noNamespaceSchemaLocation="../../../../../../UIElementsSchema/UIElements.xsd" editor-extension-mode="True">
-    <Style src="project://database/Packages/com.unity.inputsystem/InputSystem/Editor/GlobalAssetEditor/Resources/GlobalInputActionsEditorStyles.uss?fileID=7433441132597879392&amp;guid=7dac9c49a90bca4499371d0adc9b617b&amp;type=3#GlobalInputActionsEditorStyles" />
+    <Style src="project://database/Packages/com.unity.input/Runtime/Editor/GlobalAssetEditor/Resources/GlobalInputActionsEditorStyles.uss?fileID=7433441132597879392&amp;guid=7dac9c49a90bca4499371d0adc9b617b&amp;type=3#GlobalInputActionsEditorStyles" />
     <ui:VisualElement style="margin-bottom: 0; border-bottom-width: 1px; padding-bottom: 5px;">
         <ui:Foldout text="Foldout" name="Foldout" class="name-and-parameters-list-foldout" style="align-items: stretch;" />
     </ui:VisualElement>
