+++ Packages/com.unity.input/Runtime/Editor/GlobalAssetEditor/SerializedInputActionMap.cs
@@ -1,17 +1,17 @@
-﻿using System;
+using System;
 using UnityEditor;
 
 namespace UnityEngine.InputSystem.Editor
 {
-	internal readonly struct SerializedInputActionMap
-	{
-		public SerializedInputActionMap(SerializedProperty serializedProperty)
-		{
-			wrappedProperty = serializedProperty ?? throw new ArgumentNullException(nameof(serializedProperty));
-			name = serializedProperty.FindPropertyRelative(nameof(InputActionMap.m_Name)).stringValue;
-		}
+    internal readonly struct SerializedInputActionMap
+    {
+        public SerializedInputActionMap(SerializedProperty serializedProperty)
+        {
+            wrappedProperty = serializedProperty ?? throw new ArgumentNullException(nameof(serializedProperty));
+            name = serializedProperty.FindPropertyRelative(nameof(InputActionMap.m_Name)).stringValue;
+        }
 
-		public string name { get; }
-		public SerializedProperty wrappedProperty { get; }
-	}
+        public string name { get; }
+        public SerializedProperty wrappedProperty { get; }
+    }
 }
