+++ Packages/com.unity.input/Runtime/Editor/GlobalAssetEditor/Views/VisualElementExtensions.cs
@@ -1,19 +1,19 @@
-﻿using System;
+using System;
 using UnityEngine.UIElements;
 
 namespace UnityEngine.InputSystem.Editor
 {
-	public static class VisualElementExtensions
-	{
-		public static TElement Q<TElement>(this VisualElement visualElement, string name) where TElement : VisualElement
-		{
-			var element = UQueryExtensions.Q<TElement>(visualElement, name);
-			if (element == null)
-				throw new InvalidOperationException(
-					$"Expected a visual element called '{name}' of type '{typeof(TElement)}' to exist " +
-					$"but none was found.");
+    public static class VisualElementExtensions
+    {
+        public static TElement Q<TElement>(this VisualElement visualElement, string name) where TElement : VisualElement
+        {
+            var element = UQueryExtensions.Q<TElement>(visualElement, name);
+            if (element == null)
+                throw new InvalidOperationException(
+                    $"Expected a visual element called '{name}' of type '{typeof(TElement)}' to exist " +
+                    $"but none was found.");
 
-			return element;
-		}
-	}
+            return element;
+        }
+    }
 }
