+++ Packages/com.unity.input/Runtime/Editor/Settings/InputAdvancedSettingsProvider.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 18d3e83390dcd6349b3086b9f70391b9
+guid: e8df62e96c8524c2a89b7af40ee100e9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
