+++ Packages/com.unity.input/Runtime/Editor/Settings/InputSettingsProvider.cs.meta
@@ -1,11 +1,3 @@
 fileFormatVersion: 2
-guid: e8df62e96c8524c2a89b7af40ee100e9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+guid: b7b758e12c2546979d4a20c8cf493b38
+timeCreated: 1645469646
